app: remove sample config file.
Since the config.js is an actual javascript file which contains logic,
it's not really a config file at all.  If the configuration were to contain
any sensitive info, the info would be in environment variables and pulled in
by the code executing in config.js.  If we want to support any more overrides
of the settings in here, let's just make those use env vars, just as is
currently done with hostname of db service.