Fix failing publication test
This failure was a combination of two things,

the first was an issue in the angular compile / digest loop, the
solution was to put a deplay in to let angular do its buisness.

The second was an issue with the publish modal not closing until the
record was saved, thus a race condition between the save and click.