Fix failing notebook opening test.
This was a fun bug to track down. Due to the way we are loading in the
notebook there is a race condition that can happen between when you open
a notebook and when that notebook knows how to get back to the project
page.

Due to the fact that tests run quite fast, faster even than you can
click, it was exposing this race condition. The solution with this
change was to wait until the notebook title shows up on the notebook
page. At that point we can say that the notebook is ready to be "left".