Fix beaker s3 mounting.
This issue arose due to an object merge conflict between
https://github.com/twosigma/bunsen/commit/5fc067302f7ed12ed81eb12ba717903b1ec95800#diff-a4b3038d4ee7f4db64a479a6b17460a6R181
and
https://github.com/twosigma/bunsen/commit/8357bdd69ee1b2463e05f5bdac4296ce19d5a4ab#diff-ca4cafe8367c8d9733295b1cc5c50217R155

Where only the user volume scratch space was being merged into the marathon config, as compared to all of the volumes. The result of this issue was that s3 was not being mounted on each beaker container, thus users were unable to access their datasets.

The solution was to switch to a deep merge strategy between the user settings and the template settings.