docs:  Move some random tips to "quickstart" page.
The quickstart is incomplete and hints at a couple things which should be
documented more fully.  I'm not so sure we should have a "quickstart" as this
implies that one can contribute to Bunsen without really understanding the way
it works, which isn't true.  But leaving existing docs in this file for now as a
"catchall" pending more work.