tests app: refactor signup flow.
We no longer have a signup method on the user model.
There was a security issue that we were always passing in the raw params
from the request to the user forge, which made it easy for someone to
sign up with attributes that we did not want to expose to them.

Now we do a raw forge on the model via the seed route and a restricted
forge via a pluck on the sign up params via the normal auth controller.