front_end: Admin tab highlighted when children state are active
In anticipation of downgrading angular, to avoid any regression bugs, we're
going to use the angular UI router shim for active child states as we did with projects.
Angular UI router has this baked in now, but to avoid any
incompatibility we can also upgrade angular UI when we upgrade to the
latest.
https://github.com/angular-ui/ui-router/pull/927