script: Update ci url for downloading elastic search via the new domain
This was caused by an issue with elasticsearch rebranding. Their new
domain is elastic.co. All wget requests to this locally work well, but
on CI, they complain about the certificate not working. In an attempt to
squash said complain, we're switching the wget URL to be the elastic.co.