marketplace: Remove unnecessary await function from seeder
Before we were under the assumption that since `reindex-catalog!` has
agents running within it, we would have to wait for these async actions
to complete. This is not the case since all agents are taken care of
within the function(s) we're calling, via `await-for n-milliseconds`.
The `reindex-catalog!` function is overall synchronous since every async
process blocks the thread until completion.