marketplace: Rename datasets to seed-datasets
Since the purpose of this endpoint is to seed datasets in bulk and not
create an actual dataset (since it doesn't handle random ID creation),
I'm changing the endpoint and function to be prefixed with seed, so that
is more accurate. This way we can reuse the resource datasets for
creating a dataset and getting all datasets.