marketplace: Fix seeder to work on staging
The seeder only worked locally before. Environment-set URLs would never
override the defaults. `reindex-catalogs!` now takes a connection object.
Each location where this function was used is now updated to reflect this.

To seed on staging, pass an `-o` flag (for override) to the seeder. This
will override the default connection with one composed of environment
variables on the system.