marketplace: Extract related datasets from query
`find-matching` returns an map `{:data {....}}`. We were just
passing the results to the client previously. This was preventing
related datasets from showing up when viewing a single dataset, as the
client wasn't aware of the nested data. I've adjusted the data so that
it returns what the client expects and works as it previously was.