web: Temporarily hide the beaker navbar.
We will need to undo this soon in order to be able to show specific parts
of it, e.g. the menu and status feedback.  Hiding it for now because
without further design and coding it interferes with using Bunsen, being
a fixed position element across the top of the page.