web, notebook: Change open/recent Notebook sorting
Open notebooks will now be sorted by opened-at timestamp.  Therefore, the opened-at timestamp is changed to only update on the initial open, and a custom openedAt function is used to sort the timestamps since they are presented as strings and not handled correctly by angular orderBy.  Recent notebooks will also be sorted using the openedAt function for the same reason