web:  Wait to init beaker until we know the user has a beaker server.
Some parts of the beaker initialization sequence aren't sensitive to the
presence or location of the server;  those have been left in beakerEmbed.js.
The parts that require a server to be present have been moved to our existing
Beaker service whose job it is to ensure that a Beaker server is available.